Feat: gRPC (#1)
* feat: Implement gRPC client <-> server connection

- Implement the portal-server using gRPC as the protocol for the client
server communcation.

- Implement the portal-client app as a Tauri app using the gRPC
generated client code to handle the communication between client and
server.

- Implement basic UI file tree in client app

* feat: Implement `FileTree` and directory picker

* Use Catalyist for application components

- Use Catalyst sidebar layout for the main layout component

- Refactor directory open functionality to work better with the catalyst
sidebar.

* feat: Implement file opening over gRPC

Open a file on the server and stream the response back to the client.
The opened file will arrive in 1MB chunks following the suggested max
size of a bytes field for Protobuf3.

* feat: Implement file content header serde

Include file content header using custom binary protocol. This allows
us to return some extra data about the file contents along with the file
contents itself in a single `ArrayBuffer` that can be passed to JS from
the Tauri IPC.

* feat: Implement basic directory history list

* feat: Implement directory history with IndexDB

* chore: Remove unused code; fix build commands